[FEAT/#84] 캘린더 mvvm 구조 변경 및 패키지 이동
